#235 [fix] Option에서 OptionRepository 인터페이스 분리
