+++ app/src/main/java/com/gojol/notto/model/datasource/option/OptionRepository.kt
@@ -1,24 +1,12 @@
 package com.gojol.notto.model.datasource.option
 
 import com.gojol.notto.network.GithubResponse
-import javax.inject.Inject
-import javax.inject.Singleton
 
-@Singleton
-class OptionRepository @Inject constructor(
-    private val optionLocalDataSource: OptionLocalDataSource,
-    private val optionRemoteDataSource: OptionRemoteDataSource
-) : OptionDataSource {
+interface OptionRepository {
 
-    override fun isPushNotificationChecked(): Boolean {
-        return optionLocalDataSource.isPushNotificationChecked()
-    }
+    fun isPushNotificationChecked(): Boolean
 
-    override fun saveIsPushNotificationChecked(isPushChecked: Boolean) {
-        optionLocalDataSource.saveIsPushNotificationChecked(isPushChecked)
-    }
+    fun saveIsPushNotificationChecked(isPushChecked: Boolean)
 
-    override suspend fun getGitContributors(owner: String, repo: String): Result<GithubResponse> {
-        return optionRemoteDataSource.getGitContributors(owner, repo)
-    }
+    suspend fun getGitContributors(owner: String, repo: String): Result<GithubResponse>
 }
